+++ src/components/AuthPage/Login/index.js
@@ -20,78 +20,10 @@ import { Link } from "react-router-dom";
 import validator from "validator";
 import Divider from "../../../globalComponents/Divider";
 import { clearAuthError, signIn } from "../../../store/actions";
-import SmButtons from "../smButtons";
+import SmButtons from "../smButton/smButtons";
 import ViewAlerts from "./ViewAlerts";
-import { makeStyles } from "@material-ui/core/styles";
 import LoginImg from "../../../assets/images/login.svg";
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    background: "#ECEAEB",
-    height: "95vh",
-    width: "100vw",
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-    flexFlow: "row",
-  },
-  card: {
-    boxShadow: "none",
-  },
-  loginLeft: {
-    flex: "2.2",
-    [theme.breakpoints.down(750)]: {
-      flex: "0",
-    },
-  },
-  loginRight: {
-    flex: "1.8",
-    boxShadow: "none",
-    [theme.breakpoints.down(750)]: {
-      flex: "1",
-    },
-  },
-  rootChildrenLeft: {
-    flex: "1.4",
-    border: "2px solid black",
-    boxShadow: "5px 5px 10px gray",
-    background: "#759F9E",
-    animation: "$myEffectFromRight 1900ms",
-    zIndex: "2",
-  },
-  rootChildrenRight: {
-    flex: "1",
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-    flexFlow: "column",
-    animation: "$myEffectFromLeft 1500ms",
-    [theme.breakpoints.down(750)]: {
-      flex: "0",
-      display: "none",
-    },
-  },
-  "@keyframes myEffectFromRight": {
-    "0%": {
-      opacity: 1,
-      transform: "scale(.7,.7) translate(200vw) ",
-    },
-    "100%": {
-      opacity: 1,
-      transform: "scale(1,1) rotateY(0) translate(0vw)",
-    },
-  },
-  "@keyframes myEffectFromLeft": {
-    "0%": {
-      opacity: 1,
-      transform: "scale(.7,.7) translate(-100vw) ",
-    },
-    "100%": {
-      opacity: 1,
-      transform: "scale(1,1) rotateY(0) translate(0vw)",
-    },
-  },
-}));
+import useStyles from "./styles";
 
 const Login = () => {
   const firebase = useFirebase();
@@ -175,14 +107,7 @@ const Login = () => {
 
   return (
     <Grid container className={classes.root}>
-      <Grid
-        style={{
-          display: "flex",
-          justifyContent: "flex-start",
-          alignContent: "left",
-        }}
-        className={`${classes.rootChildrenLeft}`}
-      >
+      <Grid className={`${classes.rootChildrenLeft}`}>
         <Grid className={classes.loginLeft}>
           <img
             src={LoginImg}
@@ -191,11 +116,7 @@ const Login = () => {
           />
         </Grid>
 
-        <Card
-          raised
-          className={`${classes.card} ${classes.loginRight}  `}
-          // className=
-        >
+        <Card raised className={`${classes.card} ${classes.loginRight}  `}>
           <CardContent>
             <Typography
               variant="h4"
@@ -215,7 +136,6 @@ const Login = () => {
                 helperText={
                   emailValidateError ? emailValidateErrorMessage : null
                 }
-                error={emailValidateError}
                 fullWidth
                 autoComplete="email"
                 required
@@ -267,14 +187,7 @@ const Login = () => {
                 <Grid>
                   <FormGroup row>
                     <FormControlLabel
-                      control={
-                        <Checkbox
-                          // checked={state.checkedB}
-                          // onChange={handleChange}
-                          name="remember"
-                          color="primary"
-                        />
-                      }
+                      control={<Checkbox name="remember" color="primary" />}
                       label="Remember me"
                     />
                   </FormGroup>
