+++ src/components/AuthPage/SignUp/index.js
@@ -5,7 +5,7 @@ import Typography from "@material-ui/core/Typography";
 import React from "react";
 import { Link } from "react-router-dom";
 import Divider from "../../../globalComponents/Divider";
-import SmButtons from "../smButtons";
+import SmButtons from "../smButton/smButtons";
 import SignupForm from "./signupForm";
 
 const SignUp = () => {
